Set-up data for latent Dirichlet allocation

Set-up data for latent Dirichlet allocation

setUpData(ps, K, alpha = 0.9, gamma = 0.5)

Arguments

ps
phyloseq class.

K
Integer. Number of topics.

alpha
Non-negative numeric. Hyper-parameter alpha is less than one to generate sparse mixtures that are different from each other and avoid generating unrealistic topics.

gamma
Non-negative numeric. Hyper-parameter gamma is less than one to generate sparse mixtures of ASVs in each topic.

Value

List. List of object with number of topics, number of ASVs, number of specimens, count table, hyper parameters for topic and ASVs distribution.